requirements.txt
python-telegram-bot==21.6
openai-whisper==20231117
torch>=2.1.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
orjson==3.10.7
rapidfuzz==3.14.3
